+++ integration_tests/test_math.py
@@ -3,6 +3,6 @@ from modules_math import factorial
 def main0():
     x: i32
     x = 6
-    answer: i32
+    answer: i64
     answer = factorial(x)
     print(answer)
